requirements_scrapeToRag.txt
firecrawl-py>=1.0.0
python-dotenv>=1.0.0
requests>=2.30.0
